+++ mayan/apps/document_parsing/links.py
@@ -4,7 +4,8 @@ from mayan.apps.navigation.classes import Link
 from mayan.apps.navigation.utils import get_cascade_condition
 
 from .icons import (
-    icon_document_file_content, icon_document_file_content_delete,
+    icon_document_file_content, icon_document_file_content_delete_single,
+    icon_document_file_content_delete_multiple,
     icon_document_file_content_download, icon_document_file_page_content,
     icon_document_file_parsing_errors_list, icon_document_file_submit,
     icon_document_type_parsing_settings, icon_document_type_submit,
@@ -15,26 +16,28 @@ from .permissions import (
     permission_document_file_parse
 )
 
+# Document file
 
 link_document_file_content = Link(
     args='resolved_object.id', icon=icon_document_file_content,
     permissions=(permission_document_file_content_view,), text=_('Content'),
     view='document_parsing:document_file_content_view'
 )
-link_document_file_content_delete = Link(
-    args='resolved_object.id', icon=icon_document_file_content_delete,
+link_document_file_content_delete_single = Link(
+    args='resolved_object.id', icon=icon_document_file_content_delete_single,
     permissions=(permission_document_file_parse,),
     text=_('Delete parsed content'),
-    view='document_parsing:document_file_content_delete',
+    view='document_parsing:document_file_content_delete_single',
 )
-link_document_file_multiple_content_delete = Link(
-    icon=icon_document_file_content_delete, text=_('Delete parsed content'),
-    view='document_parsing:document_file_multiple_content_delete',
+link_document_file_content_delete_multiple = Link(
+    icon=icon_document_file_content_delete_multiple,
+    text=_('Delete parsed content'),
+    view='document_parsing:document_file_content_delete_multiple',
 )
 link_document_file_content_download = Link(
-    args='resolved_object.id',
-    icon=icon_document_file_content_download,
-    permissions=(permission_document_file_content_view,), text=_('Download content'),
+    args='resolved_object.id', icon=icon_document_file_content_download,
+    permissions=(permission_document_file_content_view,),
+    text=_('Download content'),
     view='document_parsing:document_file_content_download'
 )
 link_document_file_page_content = Link(
@@ -49,18 +52,21 @@ link_document_file_parsing_errors_list = Link(
     permissions=(permission_document_file_parse,), text=_('Parsing errors'),
     view='document_parsing:document_file_parsing_error_list'
 )
-link_document_file_multiple_submit = Link(
+link_document_file_submit_multiple = Link(
     icon=icon_document_file_submit,
     text=_('Submit for parsing'),
     view='document_parsing:document_file_multiple_submit'
 )
-link_document_file_submit = Link(
+link_document_file_submit_single = Link(
     args='resolved_object.id',
     icon=icon_document_file_submit,
     permissions=(permission_document_file_parse,),
     text=_('Submit for parsing'),
     view='document_parsing:document_file_submit'
 )
+
+# Document type
+
 link_document_type_parsing_settings = Link(
     args='resolved_object.id',
     icon=icon_document_type_parsing_settings,
@@ -77,6 +83,9 @@ link_document_type_submit = Link(
     text=_('Parse documents per type'),
     view='document_parsing:document_type_submit'
 )
+
+# Errors
+
 link_error_list = Link(
     icon=icon_error_list,
     permissions=(permission_document_file_parse,), text=_('Parsing errors'),